Separated E2E tests into seeded and non-seeded flows.
