+++ cypress/integration/nonSeededFlows/test.spec.js
@@ -1,0 +1,5 @@
+describe('Describe for the new E2E build node', () => {
+  it('Test for the new E2E build node', () => {
+    expect(true).toBe(true);
+  });
+});
