+++ cypress/integration/seededFlows/adminFlows/config/authenticationSection.spec.js
