+++ cypress/integration/seededFlows/adminFlows/config/getStartedSection.spec.js
