+++ cypress/integration/seededFlows/adminFlows/config/mascotSection.spec.js
