+++ cypress/integration/seededFlows/analyticsFlows/analyticsNavigation.spec.js
