+++ cypress/integration/seededFlows/articleFlows/removeMainImage.spec.js
