+++ cypress/integration/seededFlows/loginFlows/userChangePassword.spec.js
@@ -1,4 +1,4 @@
-import { getInterceptsForLingeringUserRequests } from '../../util/networkUtils';
+import { getInterceptsForLingeringUserRequests } from '../../../util/networkUtils';
 
 describe('User Change Password', () => {
   beforeEach(() => {
