+++ cypress/integration/seededFlows/navigationFlows/homePageLeftSidebar.spec.js
