+++ cypress/integration/seededFlows/notificationsFlows/notificationNavigation.spec.js
