+++ cypress/integration/seededFlows/registrationFlows/emailRegistration.spec.js
